Reordered assigned variables in constructor in the same way at outside scope Updated some variables to readonly Added check < 0 for some variables on constructor Removed unused "using"s Added null check on ScrambledKeyPair Replaced some long variables as int

+++ src/L2dotNET.GameService/Commands/Admin/AdminRange.cs
@@ -110,8 +110,8 @@ namespace L2dotNET.GameService.Commands.Admin
                 {
                     int p1X = _np._x[i];
                     int p1Y = _np._y[i];
-                    long dx = p1X - p[0],
-                         dy = p1Y - p[1];
+                    int dx = p1X - p[0],
+                        dy = p1Y - p[1];
                     double distance = Math.Sqrt((dx * dx) + (dy * dy));
                     if ((curdistance != -1) && !(distance < curdistance))
                         continue;
