+++ src/L2dotNET.GameService/network/serverpackets/TradeUpdate.cs
@@ -6,7 +6,7 @@ namespace L2dotNET.GameService.Network.Serverpackets
     class TradeUpdate : GameserverPacket
     {
         private readonly L2Item _item;
-        private readonly long _num;
+        private readonly int _num;
         private readonly byte _action;
 
         public TradeUpdate(L2Item item, int num, byte action)
